+++ blocklets/ai-kit/src/locales/zh.ts
@@ -7,7 +7,6 @@ export default flat({
   aiProviderLocalAIKit: '本地 AI Kit 组件',
   selectMonth: '选择月份',
   subscribeAITip: '现在订阅 AI Kit 来获得 AI 能力吧！',
-  subscribeAIService: '订阅 AI 服务',
   unsubscribe: '取消订阅',
   unsubscribeTip: '取消订阅后将无法继续使用我们提供的 AI 服务！',
   cancel: '取消',
@@ -19,4 +18,6 @@ export default flat({
   total: '合计',
   monthlySpend: '月度消费',
   viewSubscriptionDetail: '查看订阅详情',
+  subscriptionPastDueTip: '您的订阅已欠费，请及时支付欠款以恢复订阅服务。',
+  payNow: '立即支付',
 });